football_manager/requirements.txt
Django==1.10.3
names==0.3.0
mysql-connector==2.1.4
django-crontab==2.1.1
requests==2.13.0
